+++ ChangeLog.md
@@ -6,6 +6,7 @@
   - Add static bool VersionValid(const QString& szVersion);
   - Add static int VersionCompare(const QString &ver1, const QString &ver2);
   - Add static int GetRandomNumber(int min, int max);
+  - Add static QColor InvertColor(const QColor& color);
 - Add password generator: CPasswordGenerator
 - Add set applicaiont font
 - Add audio and video settings widget: CFrmMediaDevices